Review of 4-19-22 Updates by G. Sax
It helped to see the screenshots, which caused me to add some directional language, like "upper left" and "on the right." It may be useful to include these extra directional helpers in other guides.

It also might help to highlight the section of the report being talked about on the image itself. I'll send examples of what I mean to Jason.